Modified Sign-In button to instantly redirect to spotify sign in
