+++ jamfynder/src/App.js
@@ -7,10 +7,19 @@ import NavigationBar from "./components/NavigationBar";
 import TutorialPage from "./components/TutorialPage";
 import SettingsPage from "./components/SettingsPage";
 import AboutUs from "./components/AboutUs";
-import LoginPage from "./components/LoginPage";
 import Player from "./components/Player";
 
+const CLIENT_ID = "f12088ba0b0c45018df4dad44b51b83d";
+const ENDPOINT = "https://accounts.spotify.com/authorize";
+const REDIRECT_URL = "http://localhost:3000/Player";
+const SPACE_DELIMITER = "%20";
+const SCOPES = ["user-read-currently-playing", "user-read-playback-state"];
+const SCOPES_URL_PARM = SCOPES.join(SPACE_DELIMITER);
+
 const App = () => {
+  const handleLogin = () => {
+    window.location = `${ENDPOINT}?client_id=${CLIENT_ID}&redirect_uri=${REDIRECT_URL}&scope=${SCOPES_URL_PARM}&response_type=token&show_dialog=true`;
+  };
   return (
     <Router>
       <div className="App">
@@ -19,10 +28,6 @@ const App = () => {
           <Switch>
             <Route path="/welcome">
               <WelcomePage />
-            </Route>
-
-            <Route path="/LoginPage">
-              <LoginPage />
             </Route>
 
             <Route path="/Player">
