+++ jamfynder/src/components/NavigationBar/NavigationBar.js
@@ -2,7 +2,17 @@ import React from "react";
 import { Link, Switch, Route } from "react-router-dom";
 import "./NavigationBar.css";
 
+const CLIENT_ID = "f12088ba0b0c45018df4dad44b51b83d";
+const ENDPOINT = "https://accounts.spotify.com/authorize";
+const REDIRECT_URL = "http://localhost:3000/Player";
+const SPACE_DELIMITER = "%20";
+const SCOPES = ["user-read-currently-playing", "user-read-playback-state"];
+const SCOPES_URL_PARM = SCOPES.join(SPACE_DELIMITER);
+
 const NavigationBar = () => {
+  const handleLogin = () => {
+    window.location = `${ENDPOINT}?client_id=${CLIENT_ID}&redirect_uri=${REDIRECT_URL}&scope=${SCOPES_URL_PARM}&response_type=token&show_dialog=true`;
+  };
   return (
     <header className="NavigationBar">
       <nav>
@@ -11,7 +21,7 @@ const NavigationBar = () => {
             <li>Home</li>
           </Link>
 
-          <Link to="/LoginPage">
+          <Link onClick={handleLogin}>
             <li>Login</li>
           </Link>
