+++ jamfynder/src/components/Player/Player.css
@@ -1,8 +1,8 @@
 .wrapper .sidebar {
-  background: rgba(0, 0, 0, 0.9);
+  background: rgba(0, 0, 0, 0.75);
   position: fixed;
   top: 0;
-  left: 0;
+  right: 0;
   width: 255px;
   height: 100%;
   padding: 20px 0;
